+++ wasm-calc8/src/Calc/Typecheck.hs
@@ -3,10 +3,12 @@ module Calc.Typecheck
     module Calc.Typecheck.Helpers,
     module Calc.Typecheck.Elaborate,
     module Calc.Typecheck.Types,
+    module Calc.Typecheck.Infer,
   )
 where
 
 import Calc.Typecheck.Elaborate
 import Calc.Typecheck.Error
 import Calc.Typecheck.Helpers
+import Calc.Typecheck.Infer
 import Calc.Typecheck.Types
