+++ wasm-calc8/test/Test/Typecheck/TypecheckSpec.hs
@@ -1,38 +1,38 @@
-{-# LANGUAGE NamedFieldPuns    #-}
+{-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE OverloadedStrings #-}
 
 module Test.Typecheck.TypecheckSpec (spec) where
 
-import           Calc.ExprUtils
-import           Calc.Parser
-import           Calc.Typecheck
-import           Calc.Types.Function
-import           Calc.Types.Module
-import           Calc.Types.Op
-import           Calc.Types.Pattern
-import           Calc.Types.Type
-import           Control.Monad
-import           Data.Either         (isLeft)
-import           Data.Foldable       (traverse_)
-import qualified Data.List           as List
-import qualified Data.List.NonEmpty  as NE
-import           Data.Text           (Text)
-import           Test.Helpers
-import           Test.Hspec
+import Calc.ExprUtils
+import Calc.Parser
+import Calc.Typecheck
+import Calc.Types.Function
+import Calc.Types.Module
+import Calc.Types.Op
+import Calc.Types.Pattern
+import Calc.Types.Type
+import Control.Monad
+import Data.Either (isLeft)
+import Data.Foldable (traverse_)
+import qualified Data.List as List
+import qualified Data.List.NonEmpty as NE
+import Data.Text (Text)
+import Test.Helpers
+import Test.Hspec
 
 runTC :: TypecheckM ann a -> Either (TypeError ann) a
-runTC = runTypecheckM (TypecheckEnv mempty mempty)
+runTC = runTypecheckM (TypecheckEnv mempty mempty 0)
 
-testTypecheck :: (Text, Text) -> Spec
-testTypecheck (input, result) = it (show input) $ do
+testSucceedingExpr :: (Text, Text) -> Spec
+testSucceedingExpr (input, result) = it (show input) $ do
   case (,) <$> parseExprAndFormatError input <*> parseTypeAndFormatError result of
     Left e -> error (show e)
     Right (expr, tyResult) -> do
       getOuterAnnotation <$> runTC (infer (void expr))
         `shouldBe` Right (void tyResult)
 
-testFailing :: (Text, TypeError ()) -> Spec
-testFailing (input, result) = it (show input) $ do
+testFailingExpr :: (Text, TypeError ()) -> Spec
+testFailingExpr (input, result) = it (show input) $ do
   case parseExprAndFormatError input of
     Left e -> error (show e)
     Right expr -> do
@@ -155,6 +155,18 @@ spec = do
                     "function main() -> Int32 { acceptInt32(1) }"
                   ],
                 tyInt32
+              ),
+              ( joinLines
+                  [ "memory 1000",
+                    "function main() -> Int32 { load(0) }"
+                  ],
+                tyInt32
+              ),
+              ( joinLines
+                  [ "memory 1000",
+                    "function main() -> Int32 { store(0, (100: Int32)); load(0) }"
+                  ],
+                tyInt32
               )
             ]
       describe "Successfully typechecking modules" $ do
@@ -176,7 +188,8 @@ spec = do
               joinLines
                 [ "import console.log as log(a: Int64) -> Void",
                   "function main() -> Int32 { let a = log(1); a }"
-                ]
+                ],
+              "function noMemAllocated() -> Int32 { load(100) }"
             ]
       describe "Failing typechecking modules" $ do
         traverse_ testFailingModule failing
@@ -205,7 +218,7 @@ spec = do
             ]
 
       describe "Successfully typechecking expressions" $ do
-        traverse_ testTypecheck succeeding
+        traverse_ testSucceedingExpr succeeding
 
       let failing =
             [ ("if (1: Int64) then 1 else 2", PredicateIsNotBoolean () tyInt64),
@@ -236,4 +249,4 @@ spec = do
             ]
 
       describe "Failing typechecking expressions" $ do
-        traverse_ testFailing failing
+        traverse_ testFailingExpr failing
