+++ wasm-calc8/wasm-calc8.cabal
@@ -78,8 +78,10 @@ common shared
     Calc.Typecheck.Error
     Calc.Typecheck.Generalise
     Calc.Typecheck.Helpers
+    Calc.Typecheck.Infer
     Calc.Typecheck.Substitute
     Calc.Typecheck.Types
+    Calc.Typecheck.Unify
     Calc.Types
     Calc.Types.Annotation
     Calc.Types.Expr
